Move complex array typing out of Deserializes into HasComplexArrayTypes
We need to handle complex array types in both deserialization and
array-ification of DTOs. This moves the $complexArrayTypes property out
of Deserializes and into a new trait HasComplexArrayTypes, because
otherwise that property would be duplicated in both Deserializes and the
new trait we're going to create to handle turning DTOs into arrays.