Fix caching of schemas when normalizing
We had issues where the wrong type of arrays was being used, due to them
using the property name as the cache key. If there were two arrays called
`item` the second property would use the type of the first.

To fix this, we now cache arrays using `array-<Type>` as the cache key.

Additionally, we were caching items even if they had different
properties or description.

The caching has been extracted into it's own class to keep the logic
separate.